move tests to existing file
